refactor(modularization): Phase 4 Step 4-S3a — identity schemas extract (User + Register + Login + Auth/Config/OAuth + Invitation* + ChangePassword + UserList → aperag/domains/identity/schemas.py + dual-hook re-export) (#25)
First sub-step of Phase 4 Step 4-S3 (schemas extract, ~500 LOC canonical total across 4 domains; sub-split for CR manageability).

**New module** ``aperag/domains/identity/schemas.py``:
- 13 Pydantic schemas moved verbatim:
  ``Auth0`` / ``Authing`` / ``Logto`` / ``Auth`` / ``Config`` (auth config surface)
  + ``Register`` / ``User`` / ``Login`` / ``UserList`` / ``ChangePassword`` (user management)
  + ``InvitationCreate`` / ``Invitation`` / ``InvitationList`` (invitation flow)
- ``role`` typed as plain ``str`` / ``Literal["admin", "rw", "ro"]``
  so consumers outside identity never need the identity ``Role``
  enum (G15 canonical)
- ``_bind_view_models_reexports`` hook mirrors Phase 3 Step 4b /
  5b3 / 5a patterns — writes the 13 class objects onto
  ``aperag.schema.view_models`` via ``sys.modules`` string lookup
  (avoids tripping G1 AST scan)

**aperag/schema/view_models.py**:
- 13 class bodies removed, replaced by breadcrumb comments
- End-of-file ``try`` block extended with a second import from
  ``aperag.domains.identity.schemas`` — symmetric dual-hook pattern
  with the Phase 3 KB block
- Unused ``AnyUrl`` import removed (previously used by the
  ``FetchUrlRequest`` class carved to KB in Phase 3 Step 5a)
- ``QuotaInfo`` / ``UserQuotaInfo`` / ``UserQuotaList`` /
  ``QuotaUpdateRequest`` stay in view_models — quota's domain
  home deferred to Phase 5 / Phase 6 (msg=896584ee Section 2).

**Verification**:
- Dual-hook identity: ``from aperag.schema.view_models import User``
  resolves the same class object as the canonical
  ``aperag.domains.identity.schemas.User``.
- Full unit suite green: 671 passed / 29 skipped.
- Boundary + Phase 3 re-export gates: 17/17 green.
- ``ruff check`` + ``ruff format --check`` clean.

4-S3b (governance schemas) / 4-S3c (marketplace) / 4-S3d
(model_platform) follow.
